pyproject.toml
[project]
name = "ollama-image-captioning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=7.0.1",
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "lxml>=6.0.2",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "piexif>=1.1.3",
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
]

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
line_length = 80
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
